functionality finished
